Mission 1, 2, 3, 4 구현
