+++ src/pages/hoonjoJeong/Main/Main.scss
@@ -1,3 +1,7 @@
+.main-container {
+  margin-top: 300px;
+}
+
 .header-parents {
   display: flex;
   justify-content: center;
@@ -54,7 +58,6 @@
 .container {
   display: flex;
   justify-content: center;
-  margin-top: 120px;
 }
 
 .left-parents {
